Create ReadGenotypes.R
This is an example file on how to call the function to read genotype files into R directly. The input of the function is the file name as a string. It will output the matrix of genotypes (as integers) and a vector of IDs (strings). The individual ID should be the first column (does not have to be numeric). It assumes genotypes are integers given in a single string (no separated by spaces). No comment lines before of after or it will break for sure.